New IAP implementation, including Play support.
The old implementation was a port of the IAPv2 code to IAPv3. IAPv3 is
synchronous, while IAPv2 was not - so the result was that we had a
synchronous layer on top of an async layer on top of a sync layer.

The new Play code fits in a single class, which is quite a bit smaller.