Fix problems with Android input.
I'm pretty sure that right now, composing/completing text is simply
broken on Android. Taking a cue from SDL Hg's android-project, we
put text input into VISIBLE_PASSWORD mode, which takes the input
more directly. This seems to break languages that require input
methods.

That being said, it seems that completion and input methods are
just broken on Android at the moment - SDL2 only allows for
32 character per event, but Android really wants to work on
the entire input.

So, we're limiting things to only languages that do not require
input methods.

Beyond that, this includes a handler for performEditorAction, to
ensure input is completed and that an enter key is always passed.